test/web-platform-tests/to-upstream/xhr/content-type-unmodified.html

<!DOCTYPE html>
<title>A charset-including content-type header must not be modified (e.g. via reserialization)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

async_test(t => {
  const xhr = new XMLHttpRequest();

  xhr.open("POST", "resources/echo-content-type.py");
  xhr.setRequestHeader("content-type", "application/json; charset=UTF-8");

  xhr.onerror = t.unreached_func("Error occurred.");

  xhr.onload = t.step_func_done(() => {
    assert_equals(xhr.status, 200);
    assert_equals(xhr.responseText, "application/json; charset=UTF-8");
  });
  xhr.send("{ \"x\":\"a\"}");
});
</script>
